Only add -Werror=declaration-after-statement for 5.035004 and earlier
Perl v5.35.5 now uses some C99 features. This means that Perl's headers now
contain some code with mixed declarations and code., and so won't compile
with -Werror=declaration-after-statement

It still makes sense to add this flag for builds for earlier perl versions,
because they support long obsolete compilers that are strict in rejecting
certain C99 features, so adding this gcc flag allows us to audit that our
code does not violate this.